+++ SmartPower/Configurator.cs
@@ -14,6 +14,7 @@ using Timberborn.Localization;
 using Timberborn.MechanicalSystem;
 using Timberborn.PowerGenerating;
 using IgorZ.TimberDev.Utils.Utils;
+using Timberborn.Workshops;
 using UnityEngine.UIElements;
 
 // ReSharper disable InconsistentNaming
@@ -26,13 +27,21 @@ namespace IgorZ.SmartPower {
 sealed class Configurator : IConfigurator {
   static readonly PrefabPatcher.RequiredComponentsDep SmartAttractionDeps =
       new(typeof(Enterable), typeof(Attraction), typeof(MechanicalBuilding));
+  static readonly PrefabPatcher.RequiredComponentsDep SmartMechBuildingDeps =
+      new(typeof(Manufactory), typeof(MechanicalBuilding));
+  static readonly string PatchId = typeof(Configurator).FullName;
 
   public void Configure(IContainerDefinition containerDefinition) {
-    HarmonyPatcher.PatchRepeated(typeof(Configurator).FullName, typeof(NetworkFragmentServicePatch));
+    HarmonyPatcher.PatchRepeated(
+        PatchId,
+        typeof(NetworkFragmentServicePatch),
+        typeof(ConsumerFragmentServicePatch));
     CustomizableInstantiator.AddPatcher(
         typeof(Configurator).FullName,
         prefab => {
           PrefabPatcher.ReplaceComponent<GoodPoweredGenerator, SmartGoodPoweredGenerator>(prefab, _ => true);
+          PrefabPatcher.ReplaceComponent<MechanicalBuilding, SmartMechanicalBuilding>(
+              prefab, SmartMechBuildingDeps.Check);
           PrefabPatcher.AddComponent<SmartPoweredAttraction>(prefab, SmartAttractionDeps.Check);
         });
   }
@@ -49,6 +58,24 @@ sealed class Configurator : IConfigurator {
 
     static void Postfix(MechanicalNode mechanicalNode, Label ____label, ILoc ____loc) {
       var text = StateTextFormatter.FormatBatteryText(mechanicalNode, ____loc);
+      if (text != null) {
+        ____label.text += "\n" + text;
+      }
+    }
+  }
+
+  [HarmonyPatch]
+  static class ConsumerFragmentServicePatch {
+    const string NetworkFragmentServiceClassName = "Timberborn.MechanicalSystemUI.ConsumerFragmentService";
+    const string MethodName = "Update";
+
+    static MethodBase TargetMethod() {
+      var type = AccessTools.TypeByName(NetworkFragmentServiceClassName);
+      return AccessTools.FirstMethod(type, method => method.Name == MethodName);
+    }
+
+    static void Postfix(MechanicalNode mechanicalNode, Label ____label, ILoc ____loc) {
+      var text = StateTextFormatter.FormatBuildingText(mechanicalNode, ____loc);
       if (text != null) {
         ____label.text += "\n" + text;
       }