+++ SmartPower/package/CHANGELOG.md
@@ -1,5 +1,7 @@
 # v1.1 (pre-release)
 * [Enhancement] Add `deDe` locale file. Thanks to juf0816@Discord.
+* [Enhancement] Adjust power consumption in mauufactures when they don't produce product. In the
+  power save mode the consumption is 10% of the nominal building power.
 * [Fix] Properly update attraction state on game load.
 
 # v1.0 (12 Jul 2023)